+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Future Release
 """""""
 - Features:
     - Set the default directory of the dialog during saving a state machine in accordance with the chosen library in the library tree
+    - Create the data flows & data ports automatically in the nested states
 
 1.0.0
 """""""